% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexEffect.R
\name{flexEffect}
\alias{flexEffect}
\title{Flexible Causal Path-Specific Effect Estimation}
\usage{
flexEffect(
  p1,
  p0,
  scale = "diff",
  CI_level = 0.95,
  nboot = NULL,
  m.cores = NULL
)
}
\arguments{
\item{p1}{An object or a list of objects of class \code{flexPotential} or \code{mflexPotential} generated by the \code{flexPotential} or \code{mflexPotential} function, representing one set of potential outcomes.}

\item{p0}{An object or a list of objects of class \code{flexPotential} or \code{mflexPotential} generated by the \code{flexPotential} or \code{mflexPotential} function, representing another set of potential outcomes for comparison.
Note: \code{p1} and \code{p0} should have the same length; otherwise, one of them must contain a single potential outcome object.}

\item{scale}{A character string specifying the scale for effect estimation. Options are "diff" (difference), "risk" (risk ratio), and "oddsratio" (odds ratio). The default is "diff".}

\item{CI_level}{A numeric value indicating the confidence level for the confidence intervals, typically set between 0 and 1. The default is 0.95 for 95\% confidence intervals.}

\item{nboot}{An optional numeric value specifying the number of bootstrap samples to draw for estimating the standard errors and confidence intervals. If NULL, no bootstrapping is performed.}

\item{m.cores}{An optional numeric value indicating the number of cores to use for parallel processing during bootstrapping. If NULL, the function will use a single core.}
}
\value{
An object of class \code{flexEffect}, which includes the following components:
\describe{
\item{\code{pathsInfo}}{The original \code{pathsInfo} object used for the analysis.}
\item{\code{results}}{A data frame containing the estimated effects for each path, including the effect size, standard error, confidence intervals, and p-values (for the "diff" scale).}
\item{\code{data}}{The original data frame used for the analysis.}
\item{\code{boot_results}}{If bootstrapping is performed, a data frame containing the bootstrap estimates of the effects, including standard errors, confidence intervals, and p-values.}
\item{\code{call}}{The matched call to the \code{flexEffect} function.}
}
}
\description{
This function estimates causal path-specific effects by comparing two potential outcome objects generated from the \code{flexPotential} or \code{mflexPotential} functions.
The effects can be estimated on different scales, including difference ("diff"), risk ratio ("risk"), and odds ratio ("oddsratio").
The function also supports bootstrapping for estimating confidence intervals and p-values.
}
\examples{
# ****************************************************************************************
#    Path-specific effect for flexible pathway in one single/multiple treatment(s)
# ****************************************************************************************

###########################
# single treatment
###########################
data("singTreat")
# input information for PSE
EIF_fit <- pathsInfo(data = singTreat, A = "treat", Y = "outcome1", cov_x = c("X1", "X2"),
                     M.list = list(M1 = "med1", M2 = c('med2_1', 'med2_2'), M3 = 'med3'),
                     estimation = "EIF",
                     model.outcome = ~ glm(family = gaussian()),
                     model.treatment = ~ glm(family = binomial())
)


potential_outcome0 <- flexPotential(pathsInfo = EIF_fit, active = c(0, 0, 0, 0))
potential_outcome1 <- flexPotential(pathsInfo = EIF_fit, active = c(1, 0, 1, 1))
potential_outcome2 <- flexPotential(pathsInfo = EIF_fit, active = c(1, 0, 0, 1))
potential_outcome3 <- flexPotential(pathsInfo = EIF_fit, active = c(1, 0, 0, 0))

# one comparison
flex_results1 <- flexEffect(p1 = potential_outcome1,
                            p0 = potential_outcome0, scale = "diff", CI_level = 0.95,nboot = 100, m.cores = 3)
flex_results1

# multiple comparisons
# PSE: potential_outcome1 - potential_outcome0
# PSE: potential_outcome2 - potential_outcome3
flex_results <- flexEffect(p1 = list(potential_outcome1, potential_outcome2),
                           p0 = list(potential_outcome0, potential_outcome3), scale = "diff",
                           CI_level = 0.95,nboot = 100, m.cores = 4)
flex_results

###########################
# multiple treatments
###########################
data("multiTreat")

mfit<- pathsInfo(data = multiTreat,
                  Y = "Y",
                  A = c("t1","t2","t3"),
                  cov_x = "X",
                  M.list = list(
                    M1 = 'm1',
                    M2 = 'm2',
                    M3 = 'm3',
                    M4 = 'm4',
                    M5 = 'm5',
                    M6 = 'm6'
                  ),
                  estimation = "EIF",
                  model.treatment =list( ~  glm(family = binomial())),
                  model.outcome = list( ~SuperLearner(SL.library = "SL.mean",family = gaussian())),
                  model.iter  = list(~glm(family = gaussian()))
)

mp1<-mflexPotential(active = list(a1=c(0,1,0,0,1,0,0),
                                  a2=c(1,0,1,1,1,0,0),
                                  a3=c(NA,NA,NA,1,0,0,0)),mfit)
mp2<-mflexPotential(active = list(a1=c(0,0,0,0,1,0,0),
                                  a2=c(0,0,0,1,1,0,0),
                                  a3=c(NA,NA,NA,1,0,0,0)),mfit)

flexEffect(p1 = mp1, p0 = mp2, scale = "diff", CI_level = 0.95, nboot =100 , m.cores = 6)
}
