% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathsEffect.R
\name{pathsEffect}
\alias{pathsEffect}
\title{Causal Path-Specific Effect Estimation for Single Treatment}
\usage{
pathsEffect(
  pathsInfo,
  decomposition = "refer0",
  scale = "diff",
  CI_level = 0.95,
  nboot = NULL,
  m.cores = NULL
)
}
\arguments{
\item{pathsInfo}{An object of class \code{pathsInfo}, which is the output from the \code{pathsInfo} function, containing the necessary fitted models and data.}

\item{decomposition}{A character string specifying the method of decomposition to use for effect calculation. Options are "refer0" (aligned with CPSE) and "sequential" (aligned with NPSE). The default is "refer0".}

\item{scale}{A character string specifying the scale of the effect. Options are \code{"diff"}, \code{"risk"}, and \code{"oddsratio"}. Only the "diff" scale will provide standard errors and p-values directly from the estimated data. To obtain these for the other two scales, bootstrapping should be used.}

\item{CI_level}{A numeric value indicating the confidence level for confidence intervals, typically set between 0 and 1. The default is 0.95, which corresponds to 95\% confidence intervals.}

\item{nboot}{An optional numeric value specifying the number of bootstrap samples to draw for estimating the standard errors and confidence intervals of the effects. If NULL, no bootstrapping is performed.}

\item{m.cores}{An optional numeric value indicating the number of cores to use for parallel processing during bootstrapping. If NULL, the function will use a single core.}
}
\value{
An object of class \code{pathsEffect}, containing the following components:
\describe{
\item{\code{call}}{The matched call to the \code{pathsEffect} function.}
\item{\code{pathsInfo}}{The original \code{pathsInfo} object that was used as input.}
\item{\code{decomposition}}{A character string indicating the method of decomposition used ("refer0" or "sequential decomposition").}
\item{\code{estimation}}{The type of estimation method used, such as "EIF", "IPW", or "G-computation".}
\item{\code{scale}}{The scale used for the effect estimation, such as "diff", "risk", or "oddsratio".}
\item{\code{CI_level}}{The confidence level used for the confidence intervals.}
\item{\code{data}}{The original data set that was used for the analysis.}
\item{\code{potential_data}}{The potential outcome data generated during the analysis.}
\item{\code{results}}{A data frame containing the calculated effects for each path, including standard errors, confidence intervals, and p-values for the "diff" scale.}
\item{\code{boot_results}}{If bootstrapping is performed, a data frame containing the bootstrap estimates of the effects, including standard errors, confidence intervals, and p-values.}
}
}
\description{
This function estimates causal path-specific effects (PSE) for a single treatment,
focusing on specific pathways through which the treatment influences the outcome.
The estimation is based on the fitted models and data provided by the \code{pathsInfo} function.
}
\details{
\strong{Decomposition Methods}:
Path-specific effects can be decomposed in various ways to reveal how a treatment's effect is mediated through different pathways.
This function supports two primary methods of decomposition:
\enumerate{
\item \strong{Cumulative Path-Specific Effects (CPSE)}:
\itemize{
\item This method captures the effect of a treatment by sequentially activating all pathways that involve a particular mediator while blocking others.
\item These CPSEs can be summed to yield the total effect of the treatment, allowing the decomposition of the total effect into path-specific components. This method is particularly useful for calculating the proportion of mediation.
}
\item \strong{Natural Path-Specific Effects (NPSE)}:
\itemize{
\item NPSEs focus on the specific pathways of interest by treating the treatment variable as set to an 'active' value (A = 1) in some pathways while keeping it at a 'baseline' value (A = 0) in others. This approach helps in isolating the main effect of each mediator.
\item For example, NPSEs can be interpreted as the average change in the outcome Y if a marginalized group had the same mediator values as a privileged group.
\item Unlike CPSEs, NPSEs cannot be simply summed to represent the total effect in the presence of interactions between the treatment and mediators.
}
}
}
\examples{
# ******************************************************************************
#    Path specific effects via each mediator in one single treatment
# ******************************************************************************

data("singTreat")

# input information for PSE
EIF_fit <- pathsInfo(data = singTreat, A = "treat", Y = "outcome1", cov_x = c("X1", "X2"),
                     M.list = list(M1 = "med1", M2 = c('med2_1', 'med2_2'), M3 = 'med3'),
                     estimation = "EIF",
                     model.outcome = ~ glm(family = gaussian()),
                     model.treatment = ~ glm(family = binomial())
)

# 0 reference decomposition
results_refer0 <- pathsEffect(pathsInfo = EIF_fit, decomposition = "refer0", scale = "diff", CI_level = 0.95)
results_refer0

# sequential sequential
results_seq <- pathsEffect(pathsInfo = EIF_fit, decomposition = "sequential", scale = "diff", CI_level = 0.95)
results_seq

# bootstrap and parallel computation
results_boot <- pathsEffect(pathsInfo = EIF_fit, decomposition = "sequential",
                            scale = "diff", CI_level = 0.95, nboot = 10, m.cores = NULL)

results_parallel <- pathsEffect(pathsInfo = EIF_fit, decomposition = "sequential",
                                scale = "diff", CI_level = 0.95, nboot = 10, m.cores = 3)


}
