% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flexEffect.R
\name{flexEffect}
\alias{flexEffect}
\title{Flexible Causal Path-Specific Effect Estimation}
\usage{
flexEffect(
  p1,
  p0,
  scale = "diff",
  CI_level = 0.95,
  nboot = NULL,
  m.cores = NULL
)
}
\arguments{
\item{p1}{An object or a list of objects of class \code{flexPotential} or \code{mflexPotential} generated by the \code{flexPotential} or \code{mflexPotential} function, representing one set of potential outcomes.}

\item{p0}{An object or a list of objects of class \code{flexPotential} or \code{mflexPotential} generated by the \code{flexPotential} or \code{mflexPotential} function, representing another set of potential outcomes for comparison.
Note: \code{p1} and \code{p0} should have the same length; otherwise, one of them must contain a single potential outcome object.}

\item{scale}{A character string specifying the scale for effect estimation. Options are "diff" (difference), "risk" (risk ratio), and "oddsratio" (odds ratio). The default is "diff".}

\item{CI_level}{A numeric value indicating the confidence level for the confidence intervals, typically set between 0 and 1. The default is 0.95 for 95\% confidence intervals.}

\item{nboot}{An optional numeric value specifying the number of bootstrap samples to draw for estimating the standard errors and confidence intervals. If NULL, no bootstrapping is performed.}

\item{m.cores}{An optional numeric value indicating the number of cores to use for parallel processing during bootstrapping. If NULL, the function will use a single core.}
}
\value{
An object of class \code{flexEffect}, which includes the following components:
\describe{
\item{\code{pathsFit}}{The original \code{pathsFit} object used for the analysis.}
\item{\code{results}}{A data frame containing the estimated effects for each path, including the effect size, standard error, confidence intervals, and p-values (for the "diff" scale).}
\item{\code{data}}{The original data frame used for the analysis.}
\item{\code{boot_results}}{If bootstrapping is performed, a data frame containing the bootstrap estimates of the effects, including standard errors, confidence intervals, and p-values.}
\item{\code{call}}{The matched call to the \code{flexEffect} function.}
}
}
\description{
This function estimates causal path-specific effects by comparing two potential outcome objects generated from the \code{flexPotential} or \code{mflexPotential} functions.
The effects can be estimated on different scales, including difference ("diff"), risk ratio ("risk"), and odds ratio ("oddsratio").
The function also supports bootstrapping for estimating confidence intervals and p-values.
}
\examples{
# Example usage
data("singTreat")

fit <- pathsFit(data = singTreat, A = "treat", Y = "outcome1", cov_x = c("X1", "X2"),
                M.list = list(M1 = "med1", M2 = c('med2_1', 'med2_2'), M3 = 'med3'),
                estimation = "EIF",
                model.outcome = list(~ glm(family = gaussian())),
                model.propensity = ~ bart(verbose = FALSE, ndpost = 200))

potential_outcome1 <- flexPotential( pathsFit = fit, active = c(1, 0, 1, 1))
potential_outcome2 <- flexPotential( pathsFit = fit, active = c(0, 1, 0, 1))
potential_outcome0 <- flexPotential( pathsFit = fit, active = c(0, 0, 0, 0))

effect_results <- flexEffect(p1 = list(potential_outcome1, potential_outcome2),
                             p0 = potential_outcome0, scale = "diff", CI_level = 0.95)

}
